+++ src/config.js
@@ -1,9 +1,7 @@
-import variants from "./variants.js";
 import theme from "./theme.js";
 import vars from "./vars.js";
 
 const configOptions = {
-  variants,
   theme,
   vars,
 };